+++ lib/IssuanceInitiation.ts
@@ -1,8 +1,8 @@
 import { convertJsonToURI, convertURIToJsonObject } from './functions';
 import { IssuanceInitiationRequestPayload, IssuanceInitiationWithBaseUrl } from './types';
 
-export default class IssuanceInitiation {
-  static fromURI(issuanceInitiationURI: string): IssuanceInitiationWithBaseUrl {
+export class IssuanceInitiation {
+  public static fromURI(issuanceInitiationURI: string): IssuanceInitiationWithBaseUrl {
     if (!issuanceInitiationURI.includes('?')) {
       throw new Error('Invalid Issuance Initiation Request Payload');
     }
@@ -18,7 +18,7 @@ export default class IssuanceInitiation {
     };
   }
 
-  static toURI(issuanceInitiation: IssuanceInitiationWithBaseUrl): string {
+  public static toURI(issuanceInitiation: IssuanceInitiationWithBaseUrl): string {
     return convertJsonToURI(issuanceInitiation.issuanceInitiationRequest, {
       baseUrl: issuanceInitiation.baseUrl,
       arrayTypeProperties: ['credential_type'],
