+++ lib/types/Generic.types.ts
@@ -1,0 +1,6 @@
+export interface ErrorResponse extends Response {
+  error: string;
+  error_description?: string;
+  error_uri?: string;
+  state?: string;
+}
